Loads of cool stuff
Many, many fixes...
No more compiler warnings!
Surprises.... ;)
The GUI is in the process of being rewritten so it's all broken!

+++ Engine/Font.cpp
@@ -41,7 +41,7 @@ void Font::init(FT_Library& lib,const char* path, int size) {
 			texture,
 			glm::ivec2(_face->glyph->bitmap.width, _face->glyph->bitmap.rows),
 			glm::ivec2(_face->glyph->bitmap_left, _face->glyph->bitmap_top),
-			_face->glyph->advance.x
+			(GLuint)_face->glyph->advance.x
 		};
 		_chars.insert(std::pair<GLchar, Character>(i, character));
 	}
